Original Commit.
This package is in a beta version. It's still under testing and developing, one uses it at one's risk.

SPMF is a java library developped by Philippe Fournier Viger in order to mine frequent itemsets and sequences. The input format is frugal but not very intuitive. This package provides functions to convert dataframes into input files, and functions to query SPMF algorithms.
Some functions helps the building of the transaction database in, case the ID of the customer does not define the sequence to be build.